backend/requirements.txt
opencv-python
numpy
matplotlib
scipy
pdf2image
pytesseract
Pillow
openai
python-dotenv
flask
flask-cors
flask-login
flask-dance
requests
werkzeug
gunicorn
PyMuPDF==1.26.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
google-cloud-storage==2.10.0
google-cloud-pubsub==2.18.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
alembic==1.12.1
psutil
pytest
